+++ Cargo.toml
@@ -3,9 +3,6 @@ name = "rust-os"
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-crate-type = ["staticlib"]
-
 [dependencies]
 lazy_static = { version = "1", features = ["spin_no_std"] }
 limine = "0.1.10"
