+++ src/lib.rs
@@ -1,8 +1,6 @@
 #![no_std]
 #![feature(abi_x86_interrupt)]
 
-mod gdt;
-mod interrupts;
-mod kmain;
-mod serial;
-mod vga_buffer;
+pub mod gdt;
+pub mod interrupts;
+pub mod serial;
